+++ src/audio_input/lowpassfilter.cpp
@@ -3,16 +3,16 @@
 #include <stdexcept>
 
 
-LowPassFilter::LowPassFilter(double const * _filterConsts, const int _filterLength, const int _dataLength) : filterLength(_filterLength), dataLength(_dataLength) {
+LowPassFilter::LowPassFilter(float const * _filterConsts, const int _filterLength, const int _dataLength) : filterLength(_filterLength), dataLength(_dataLength) {
     filterConsts = _filterConsts;
-    oldData = new double[filterLength]();
+    oldData = new float[filterLength]();
 }
 
 LowPassFilter::~LowPassFilter() {
     delete[] oldData;
 }
 
-void LowPassFilter::filter(double const * input, double * output) {
+void LowPassFilter::filter(float const * input, float * output) {
 
     //Filter input that use old data
     for(int i=0;i < filterLength-1;i++) {
@@ -34,13 +34,13 @@ void LowPassFilter::filter(double const * input, double * output) {
     }
 
     //Copy new data to old data
-    memcpy(oldData, input+dataLength-(filterLength-1), (filterLength-1)*sizeof(double));
+    memcpy(oldData, input+dataLength-(filterLength-1), (filterLength-1)*sizeof(float));
 
     return;
 
 }
 
-void LowPassFilter::downsample(const double *input, double *output, int dataLen, int downsampleFactor)
+void LowPassFilter::downsample(const float *input, float *output, int dataLen, int downsampleFactor)
 {
     if (downsampleFactor < 2) {
         throw std::invalid_argument("Downsampling factor is not greater than 1!");
